+++ docs/_includes/projects/docs/templates/compose/my-page.js
@@ -8,13 +8,13 @@ class MyPage extends LitElement {
       ${this.footerTemplate}
     `;
   }
-  static get headerTemplate() {
+  get headerTemplate() {
     return html`<header>header</header>`;
   }
-  static get articleTemplate() {
+  get articleTemplate() {
     return html`<article>article</article>`;
   }
-  static get footerTemplate() {
+  get footerTemplate() {
     return html`<footer>footer</footer>`;
   }
 }