+++ docs/_includes/projects/docs/templates/renderroot/light-dom.js
@@ -1,19 +1,20 @@
 import { LitElement, html } from 'lit-element';
 
 /**
- * This element renders its template as light DOM children.
+ * This element renders its template contents as children, instead of 
+ * rendering into a shadow tree.
  */
 class LightDom extends LitElement {
   render(){
     return html`
-      <p><b>Render root overridden.</b> Template renders in light DOM.</p>
+      <p><b>Render root overridden.</b> Template renders without shadow DOM.</p>
     `;
   }
   /**
    * To customize an element's render root, implement createRenderRoot. Return
    * the node into which to render the element's template.
    *
-   * This element renders child nodes into its light DOM.
+   * This element renders without shadow DOM.
    */
   createRenderRoot(){
     return this;