+++ docs/_includes/projects/events/events/my-element.js
@@ -1,7 +1,7 @@
 import { LitElement, html } from '@polymer/lit-element';
 
-class MyElement extends LitElement {  
-  static get properties() { return { 
+class MyElement extends LitElement {
+  static get properties() { return {
     prop1: { type: String },
     prop2: { type: Number },
     prop3: { type: Boolean },
@@ -23,7 +23,7 @@ class MyElement extends LitElement {
       <p>prop1: ${this.prop1}</p>
       <p>prop2: ${this.prop2}</p>
       <p>prop3: ${this.prop3}</p>
-      <p>prop4[0]:</p>${this.prop4[0]}</p>
+      <p>prop4[0]: ${this.prop4[0]}</p>
       <p>prop5.subprop1: ${this.prop5.subprop1}</p>
     `;
   }