+++ src/lib/updating-element.ts
@@ -127,7 +127,12 @@ type PropertyDeclarationMap = Map<PropertyKey, PropertyDeclaration>;
 
 type AttributeMap = Map<string, PropertyKey>;
 
-export type PropertyValues = Map<PropertyKey, unknown>;
+/**
+ * Map of changed properties with old values. Takes an optional generic
+ * interface corresponding to the declared element properties.
+ */
+// tslint:disable-next-line:no-any
+export type PropertyValues<T = any> = keyof T extends PropertyKey ? Map<keyof T, unknown> : never;
 
 export const defaultConverter: ComplexAttributeConverter = {
 
@@ -180,16 +185,21 @@ const defaultPropertyDeclaration: PropertyDeclaration = {
   hasChanged: notEqual
 };
 
-const microtaskPromise = Promise.resolve(true);
-
 const STATE_HAS_UPDATED = 1;
 const STATE_UPDATE_REQUESTED = 1 << 2;
 const STATE_IS_REFLECTING_TO_ATTRIBUTE = 1 << 3;
 const STATE_IS_REFLECTING_TO_PROPERTY = 1 << 4;
-const STATE_HAS_CONNECTED = 1 << 5;
 type UpdateState = typeof STATE_HAS_UPDATED|typeof STATE_UPDATE_REQUESTED|
     typeof STATE_IS_REFLECTING_TO_ATTRIBUTE|
-    typeof STATE_IS_REFLECTING_TO_PROPERTY|typeof STATE_HAS_CONNECTED;
+    typeof STATE_IS_REFLECTING_TO_PROPERTY;
+
+/**
+ * The Closure JS Compiler doesn't currently have good support for static
+ * property semantics where "this" is dynamic (e.g.
+ * https://github.com/google/closure-compiler/issues/3177 and others) so we use
+ * this hack to bypass any rewriting by the compiler.
+ */
+const finalized = 'finalized';
 
 /**
  * Base element class which manages element properties and attributes. When
@@ -213,7 +223,7 @@ export abstract class UpdatingElement extends HTMLElement {
   /**
    * Marks class as having finished creating properties.
    */
-  protected static finalized = true;
+  protected static[finalized] = true;
 
   /**
    * Memoized list of all class properties, including any superclass properties.
@@ -296,15 +306,13 @@ export abstract class UpdatingElement extends HTMLElement {
     Object.defineProperty(this.prototype, name, {
       // tslint:disable-next-line:no-any no symbol in index
       get(): any {
-        // tslint:disable-next-line:no-any no symbol in index
-        return (this as any)[key];
+        return (this as {[key: string]: unknown})[key as string];
       },
       set(this: UpdatingElement, value: unknown) {
-        // tslint:disable-next-line:no-any no symbol in index
-        const oldValue = (this as any)[name];
-        // tslint:disable-next-line:no-any no symbol in index
-        (this as any)[key] = value;
-        this.requestUpdate(name, oldValue);
+        const oldValue =
+            (this as {} as {[key: string]: unknown})[name as string];
+        (this as {} as {[key: string]: unknown})[key as string] = value;
+        (this as unknown as UpdatingElement)._requestUpdate(name, oldValue);
       },
       configurable: true,
       enumerable: true
@@ -317,16 +325,12 @@ export abstract class UpdatingElement extends HTMLElement {
    * @nocollapse
    */
   protected static finalize() {
-    if (this.hasOwnProperty(JSCompiler_renameProperty('finalized', this)) &&
-        this.finalized) {
-      return;
-    }
     // finalize any superclasses
     const superCtor = Object.getPrototypeOf(this);
-    if (typeof superCtor.finalize === 'function') {
+    if (!superCtor.hasOwnProperty(finalized)) {
       superCtor.finalize();
     }
-    this.finalized = true;
+    this[finalized] = true;
     this._ensureClassProperties();
     // initialize Map populated in observedAttributes
     this._attributeToPropertyMap = new Map();
@@ -416,8 +420,11 @@ export abstract class UpdatingElement extends HTMLElement {
 
   private _updateState: UpdateState = 0;
   private _instanceProperties: PropertyValues|undefined = undefined;
-  private _updatePromise: Promise<unknown> = microtaskPromise;
-  private _hasConnectedResolver: (() => void)|undefined = undefined;
+  // Initialize to an unresolved Promise so we can make sure the element has
+  // connected before first update.
+  private _updatePromise =
+      new Promise((res) => this._enableUpdatingResolver = res);
+  private _enableUpdatingResolver: (() => void)|undefined;
 
   /**
    * Map with keys for any properties that have changed since the last
@@ -442,6 +449,9 @@ export abstract class UpdatingElement extends HTMLElement {
    */
   protected initialize() {
     this._saveInstanceProperties();
+    // ensures first update will be caught by an early access of
+    // `updateComplete`
+    this._requestUpdate();
   }
 
   /**
@@ -484,16 +494,15 @@ export abstract class UpdatingElement extends HTMLElement {
   }
 
   connectedCallback() {
-    this._updateState = this._updateState | STATE_HAS_CONNECTED;
-    // Ensure connection triggers an update. Updates cannot complete before
-    // connection and if one is pending connection the `_hasConnectionResolver`
-    // will exist. If so, resolve it to complete the update, otherwise
-    // requestUpdate.
-    if (this._hasConnectedResolver) {
-      this._hasConnectedResolver();
-      this._hasConnectedResolver = undefined;
-    } else {
-      this.requestUpdate();
+    // Ensure first connection completes an update. Updates cannot complete
+    // before connection.
+    this.enableUpdating();
+  }
+
+  protected enableUpdating() {
+    if (this._enableUpdatingResolver !== undefined) {
+      this._enableUpdatingResolver();
+      this._enableUpdatingResolver = undefined;
     }
   }
 
@@ -566,6 +575,44 @@ export abstract class UpdatingElement extends HTMLElement {
   }
 
   /**
+   * This private version of `requestUpdate` does not access or return the
+   * `updateComplete` promise. This promise can be overridden and is therefore
+   * not free to access.
+   */
+  private _requestUpdate(name?: PropertyKey, oldValue?: unknown) {
+    let shouldRequestUpdate = true;
+    // If we have a property key, perform property update steps.
+    if (name !== undefined) {
+      const ctor = this.constructor as typeof UpdatingElement;
+      const options =
+          ctor._classProperties!.get(name) || defaultPropertyDeclaration;
+      if (ctor._valueHasChanged(
+              this[name as keyof this], oldValue, options.hasChanged)) {
+        if (!this._changedProperties.has(name)) {
+          this._changedProperties.set(name, oldValue);
+        }
+        // Add to reflecting properties set.
+        // Note, it's important that every change has a chance to add the
+        // property to `_reflectingProperties`. This ensures setting
+        // attribute + property reflects correctly.
+        if (options.reflect === true &&
+            !(this._updateState & STATE_IS_REFLECTING_TO_PROPERTY)) {
+          if (this._reflectingProperties === undefined) {
+            this._reflectingProperties = new Map();
+          }
+          this._reflectingProperties.set(name, options);
+        }
+      } else {
+        // Abort the request if the property should not be considered changed.
+        shouldRequestUpdate = false;
+      }
+    }
+    if (!this._hasRequestedUpdate && shouldRequestUpdate) {
+      this._updatePromise = this._enqueueUpdate();
+    }
+  }
+
+  /**
    * Requests an update which is processed asynchronously. This should
    * be called when an element should update based on some state not triggered
    * by setting a property. In this case, pass no arguments. It should also be
@@ -579,32 +626,7 @@ export abstract class UpdatingElement extends HTMLElement {
    * @returns {Promise} A Promise that is resolved when the update completes.
    */
   requestUpdate(name?: PropertyKey, oldValue?: unknown) {
-    let shouldRequestUpdate = true;
-    // if we have a property key, perform property update steps.
-    if (name !== undefined && !this._changedProperties.has(name)) {
-      const ctor = this.constructor as typeof UpdatingElement;
-      const options =
-          ctor._classProperties!.get(name) || defaultPropertyDeclaration;
-      if (ctor._valueHasChanged(
-              this[name as keyof this], oldValue, options.hasChanged)) {
-        // track old value when changing.
-        this._changedProperties.set(name, oldValue);
-        // add to reflecting properties set
-        if (options.reflect === true &&
-            !(this._updateState & STATE_IS_REFLECTING_TO_PROPERTY)) {
-          if (this._reflectingProperties === undefined) {
-            this._reflectingProperties = new Map();
-          }
-          this._reflectingProperties.set(name, options);
-        }
-        // abort the request if the property should not be considered changed.
-      } else {
-        shouldRequestUpdate = false;
-      }
-    }
-    if (!this._hasRequestedUpdate && shouldRequestUpdate) {
-      this._enqueueUpdate();
-    }
+    this._requestUpdate(name, oldValue);
     return this.updateComplete;
   }
 
@@ -612,31 +634,23 @@ export abstract class UpdatingElement extends HTMLElement {
    * Sets up the element to asynchronously update.
    */
   private async _enqueueUpdate() {
-    // Mark state updating...
     this._updateState = this._updateState | STATE_UPDATE_REQUESTED;
-    let resolve: (r: boolean) => void;
-    const previousUpdatePromise = this._updatePromise;
-    this._updatePromise = new Promise((res) => resolve = res);
-    // Ensure any previous update has resolved before updating.
-    // This `await` also ensures that property changes are batched.
-    await previousUpdatePromise;
-    // Make sure the element has connected before updating.
-    if (!this._hasConnected) {
-      await new Promise((res) => this._hasConnectedResolver = res);
+    try {
+      // Ensure any previous update has resolved before updating.
+      // This `await` also ensures that property changes are batched.
+      await this._updatePromise;
+    } catch (e) {
+      // Ignore any previous errors. We only care that the previous cycle is
+      // done. Any error should have been handled in the previous update.
     }
-    // Allow `performUpdate` to be asynchronous to enable scheduling of updates.
     const result = this.performUpdate();
-    // Note, this is to avoid delaying an additional microtask unless we need
-    // to.
-    if (result != null &&
-        typeof (result as PromiseLike<unknown>).then === 'function') {
+    // If `performUpdate` returns a Promise, we await it. This is done to
+    // enable coordinating updates with a scheduler. Note, the result is
+    // checked to avoid delaying an additional microtask unless we need to.
+    if (result != null) {
       await result;
     }
-    resolve!(!this._hasRequestedUpdate);
-  }
-
-  private get _hasConnected() {
-    return (this._updateState & STATE_HAS_CONNECTED);
+    return !this._hasRequestedUpdate;
   }
 
   private get _hasRequestedUpdate() {
@@ -648,10 +662,13 @@ export abstract class UpdatingElement extends HTMLElement {
   }
 
   /**
-   * Performs an element update.
+   * Performs an element update. Note, if an exception is thrown during the
+   * update, `firstUpdated` and `updated` will not be called.
    *
-   * You can override this method to change the timing of updates. For instance,
-   * to schedule updates to occur just before the next frame:
+   * You can override this method to change the timing of updates. If this
+   * method is overridden, `super.performUpdate()` must be called.
+   *
+   * For instance, to schedule updates to occur just before the next frame:
    *
    * ```
    * protected async performUpdate(): Promise<unknown> {
@@ -665,16 +682,28 @@ export abstract class UpdatingElement extends HTMLElement {
     if (this._instanceProperties) {
       this._applyInstanceProperties();
     }
-    if (this.shouldUpdate(this._changedProperties)) {
-      const changedProperties = this._changedProperties;
-      this.update(changedProperties);
+    let shouldUpdate = false;
+    const changedProperties = this._changedProperties;
+    try {
+      shouldUpdate = this.shouldUpdate(changedProperties);
+      if (shouldUpdate) {
+        this.update(changedProperties);
+      }
+    } catch (e) {
+      // Prevent `firstUpdated` and `updated` from running when there's an
+      // update exception.
+      shouldUpdate = false;
+      throw e;
+    } finally {
+      // Ensure element can accept additional updates after an exception.
+      this._markUpdated();
+    }
+    if (shouldUpdate) {
       if (!(this._updateState & STATE_HAS_UPDATED)) {
         this._updateState = this._updateState | STATE_HAS_UPDATED;
         this.firstUpdated(changedProperties);
       }
       this.updated(changedProperties);
-    } else {
-      this._markUpdated();
     }
   }
 
@@ -687,15 +716,38 @@ export abstract class UpdatingElement extends HTMLElement {
    * Returns a Promise that resolves when the element has completed updating.
    * The Promise value is a boolean that is `true` if the element completed the
    * update without triggering another update. The Promise result is `false` if
-   * a property was set inside `updated()`. This getter can be implemented to
-   * await additional state. For example, it is sometimes useful to await a
-   * rendered element before fulfilling this Promise. To do this, first await
-   * `super.updateComplete` then any subsequent state.
+   * a property was set inside `updated()`. If the Promise is rejected, an
+   * exception was thrown during the update.
+   *
+   * To await additional asynchronous work, override the `_getUpdateComplete`
+   * method. For example, it is sometimes useful to await a rendered element
+   * before fulfilling this Promise. To do this, first await
+   * `super._getUpdateComplete()`, then any subsequent state.
    *
    * @returns {Promise} The Promise returns a boolean that indicates if the
    * update resolved without triggering another update.
    */
   get updateComplete() {
+    return this._getUpdateComplete();
+  }
+
+  /**
+   * Override point for the `updateComplete` promise.
+   *
+   * It is not safe to override the `updateComplete` getter directly due to a
+   * limitation in TypeScript which means it is not possible to call a
+   * superclass getter (e.g. `super.updateComplete.then(...)`) when the target
+   * language is ES5 (https://github.com/microsoft/TypeScript/issues/338).
+   * This method should be overridden instead. For example:
+   *
+   *   class MyElement extends LitElement {
+   *     async _getUpdateComplete() {
+   *       await super._getUpdateComplete();
+   *       await this._myChild.updateComplete;
+   *     }
+   *   }
+   */
+  protected _getUpdateComplete() {
     return this._updatePromise;
   }
 
@@ -704,7 +756,7 @@ export abstract class UpdatingElement extends HTMLElement {
    * an update. By default, this method always returns `true`, but this can be
    * customized to control when to update.
    *
-   * * @param _changedProperties Map of changed properties with old values
+   * @param _changedProperties Map of changed properties with old values
    */
   protected shouldUpdate(_changedProperties: PropertyValues): boolean {
     return true;
@@ -716,7 +768,7 @@ export abstract class UpdatingElement extends HTMLElement {
    * Setting properties inside this method will *not* trigger
    * another update.
    *
-   * * @param _changedProperties Map of changed properties with old values
+   * @param _changedProperties Map of changed properties with old values
    */
   protected update(_changedProperties: PropertyValues) {
     if (this._reflectingProperties !== undefined &&
@@ -737,7 +789,7 @@ export abstract class UpdatingElement extends HTMLElement {
    * Setting properties inside this method will trigger the element to update
    * again after this update cycle completes.
    *
-   * * @param _changedProperties Map of changed properties with old values
+   * @param _changedProperties Map of changed properties with old values
    */
   protected updated(_changedProperties: PropertyValues) {
   }
@@ -749,7 +801,7 @@ export abstract class UpdatingElement extends HTMLElement {
    * Setting properties inside this method will trigger the element to update
    * again after this update cycle completes.
    *
-   * * @param _changedProperties Map of changed properties with old values
+   * @param _changedProperties Map of changed properties with old values
    */
   protected firstUpdated(_changedProperties: PropertyValues) {
   }